src/app/requirements.txt
# Flask Web框架
Flask==3.0.0
Werkzeug==3.0.1

# 深度学习
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78

# 模型相关
# 注意：torchvision会自动安装torch依赖
